% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{dupsBetweenDFs}
\alias{dupsBetweenDFs}
\title{Find rows that are duplicated across dataframes}
\usage{
dupsBetweenDFs(dfs, by = NULL, matchall = FALSE)
}
\arguments{
\item{dfs}{a list of dataframes to compare}

\item{by}{a list of vectors each of which indicates which columns to use for
the corresponding dataframe in 'dfs'. A NULL value means use all columns.}

\item{matchall}{If TRUE then find rows that are duplicated across ALL groups, 
otherwise a row only need be duplicated across 2 groups}
}
\value{
a list of vectors indicating duplicated rows of dataframes
}
\description{
Find rows that are duplicated across dataframes
}
\details{
This is just a wrapper around \code{\link{dupsBetweenGroups}} (which see),
but for separate dataframes instead of groups within a single dataframe.

The 'by' argument can be used for specifying which columns to restrict each
dataframe in 'dfs' to. Each element of 'by' can be either NULL (default) or
a character/numeric vector. A NULL value means use all columns and a character
or numeric vector means use the columns indicated by that vector.
}
\author{
Ben Veal
}
