\name{is.logical.contents}
\alias{is.logical.contents}
\title{Test if contents of vector are logical (TRUE/FALSE) values.}
\usage{
  is.logical.contents(x)
}
\arguments{
  \item{x}{}
}
\value{
  TRUE or FALSE
}
\description{
  Test if contents of vector are logical (TRUE/FALSE)
  values.
}
\details{
  Unlike \code{\link{is.character}} this works with
  factors.
}
\author{
  Ben Veal
}
\seealso{
  \code{\link{contents}},
  \code{\link{is.character.contents}},
  \code{\link{is.numeric.contents}},
}

