% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{numUnique}
\alias{numUnique}
\title{Return the number of unique elements of x (excluding NA & NaN's)}
\usage{
numUnique(x, warn = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{warn}{Whether to issue warnings from 'unique' (default is FALSE).}
}
\value{
The number of unique values of 'x' excluding NA & NaN values.
}
\description{
Return the number of unique elements of x (excluding NA & NaN's)
}
\author{
Ben Veal
}
\keyword{misc}
\keyword{utilities}
