% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{unfactor}
\alias{unfactor}
\title{Convert factors into numeric/character/logical vectors.}
\usage{
unfactor(x)
}
\arguments{
\item{x}{A factor vector, list or dataframe.}
}
\value{
A vector, list or dataframe containing no factors.
}
\details{
This function converts factor vectors into a numeric, character or logical vectors
depending on the contents. If 'x' is a list or dataframe then any constituent factors will
be converted.
}
\author{
Ben Veal
}
