% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{is.character.contents}
\alias{is.character.contents}
\title{Test if contents of vector or factor are character/string values.}
\usage{
is.character.contents(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE or FALSE
}
\description{
Test if contents of vector or factor are character/string values.
}
\details{
Unlike \code{\link{is.character}} this works with factors.
}
\author{
Ben Veal
}
\seealso{
\code{\link{contents}}, \code{\link{is.numeric.contents}}, \code{\link{is.logical.contents}}
}

