% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_utils.R
\name{is.logical.contents}
\alias{is.logical.contents}
\title{Test if contents of vector or factor are logical (TRUE/FALSE) values.}
\usage{
is.logical.contents(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE or FALSE
}
\details{
Unlike \code{\link{is.character}} this works with factors.
}
\seealso{
\code{\link{contents}}, \code{\link{is.character.contents}}, \code{\link{is.numeric.contents}},
}
\author{
Ben Veal
}
